Update MathSearch tests
Uses the new npm test

Adds composer test. ( I Will do a follow up with php code sniffer )

Add test for js for when users add js for this extension. Currently there
is no js in this extension but just in case someone adds it this test will
be waiting for it.

Update qqq.json message.

Required-by: https://gerrit.wikimedia.org/r/#/c/228277/

Change-Id: Ib530ea6ae0dfe3ee48cc544508ce7522007e3961