Don't configure parser-factory-callback
Configuring parser-factory-callback is not necessary here because it's
the same as the default for the string value type. Setting a data type
specific custom parser will soon allow for data type specific overrides
of the parser whenever a statement value is deserialized, and setting
it may result in additional data type lookups.

Bug: T360603
Change-Id: I96fc9bb8053e7bd995ab72d4b9694f85055a98bf