Make use of modern PHP syntax like ??= or str_contains
All this code should do the exact same as before, as proven by the
tests.

Also: The [] operator to append an element to an array auto-initializes
deeper levels of the array if needed, as long as the first level is
already an array. No need for the extra manual initialization.

Change-Id: I7680d13a61a4bfd7eb23eb3bc54dc8540d0c05e5