Further improve more complex PHPDocs for search results
See I885ad2b for the previous discussion. The problem here is that
the search results in the MathSearchTerm class can be either an array
of SearchResult objects, or a trivial nested array with only scalar
values. This depends on the type of search and is hard to predict and
hard to document.

Change-Id: Ic3b6ab47457a829889165c958e866409848fd79e