MathSearchHooks: Replace hook with MathFormulaPostRenderRevision
In Math extension, original hook MathFormulaPostRender was sending
legacy parser instance. With parsoid, this instance is not available,
so the new version of the hook is needed.
MathFormulaPostRenderRevision hook provides the same function as
the MathFormulaPostRender, but instead of Parser, RevisionRecord
is being sent.

This patch provides replacement from the old hook to the new one.

Bug: T399226
Depends-On: Ie1def0b7cbb8cf928a44bb5e8431889049faf813
Change-Id: I4d36f4ea2c25f173a0be7039f69c33d98ed29733